refactor(stack): split inventory rule model into private rules.rs (GH-1731 T6)
Behavior-neutral module split that makes room for the six post-merge
remediation findings (SP1731-T7..T9). inventory.rs becomes an inventory/
directory module; the rule model (DirSelector, selectors, RuleTarget,
Matcher, StaticRule, STATIC_RULES, config shapes, configured-source
normalization) moves into a private rules.rs. Traversal helpers and the
public contract stay in mod.rs. No public API, serde shape, rule order,
or behavior change.

mod.rs 629 lines, rules.rs 184 lines — both under the 800-line ceiling.

Verified: cargo test -p harness-core stack::inventory -> 25 passed;
clippy --all-targets -D warnings clean; fmt clean.

Refs #1731
